Use synchronous connections
Connections are now returned once they've connected to the bus.

We also trap exits on the connection process and sleep a bit to try and allow supervisor-initiated restarts to be of some use.

init now also obeys the contract of init functions, which should return `:stop` tuples instead of `:error`  tuples in failure cases.