Simplify foreign command execution
Sometimes the simplest way is the best way. I realized `Porcelain.exec`
did everything we need if we were careful about how we construct the
proplist used for the `env` option.

This neatly fixes the problems I had w/goon panicking and allows me to
remove a bunch of code I'd written to handle potential races reading
stdout, stderr, and terminating the foreign command's process. The only
drawback to this approach is we have no control over how long the
foreign command will run. We can iterate on this design further as
needed and look to `Porcelain.spawn/3` and/or `Porcelain.spawn_shell/2`
for ideas on enforcing timeouts.