Allow Elixir commands to formally signal errors
Previously, all command invocations, regardless of whether or not they
actually succeeded, were reported back to the executor as successful.

Now, we can signal invocation failures by returning an `error` tuple,
containing an error message string that will ultimately be reported back
to the chat user.

The only existing call site of the `GenCommand.send_error_reply/3`
function was where an error is reported due to a failed decoding of a
Request struct. This returns a map and not a string. Long term there is
thought of more full-featured error handling, including HTTP-like status
codes, messages, and more complex bodies, perhaps even with error
templating. For now, though, we'll start with a simple
error-message-only interface. As a result, that error map is now turned
into a string and used as the error message.

Note that the Cog executor already has code in place to handle this;
existing commands will need to be altered to return the appropriate
tuples.