Define new template specification
Previously, bundles could include both Slack and HipChat templates. With
the new templating engine, we now only need one template (the specific
chat provider implementations will be responsible for rendering the
template in a manner appropriate to the chat platform).

Now, templates look like this:

```
...
"templates": {
  "template1": {"body": "template1 body"},
  "template2": {"body": "template2 body"},
  ...
}
...

```

Note now that template names are also constrained to the regex
`[A-Za-z0-9_]+`.

This sets the current bundle definition version to 4, and officially
removes support for version 2 bundles. Version 3 is now officially
deprecated.