Bump configuration version to 5
This officially removes support for v3 bundles (the last to support
Mustache templates).

There are currently no structural changes that differentiate v4
bundles from v5; this increment is more a reflection of a change in
the execution environment these bundles are meant to run in. We don't
currently have a formal mechanism to represent such environmental
changes. Until such time as we do, the bundle configuration version
should be understood to encompass both structural features of the
configuration file _and_ their execution environment.

(Note, however, that v5 will not be finalized until Cog 1.0 is
released in mid January 2017; structural changes to the configuration
file schema may be added between now and then.)

Some tests for v4 were updated so they better capture the behavior being
tested (that is, we're testing that users are urged to update to the
latest valid configuration version).